FastAPI/requirements.txt
fastapi
uvicorn
sentence-transformers
faiss-cpu
huggingface-hub
numpy
pandas
gdown
python-dotenv
tqdm
torch
